vmax: possible vmax stuck fixed
<==== Mercurial data ====>
Branch: stable

+++ common/src/plugins/storage/dts/vmax480/vmax480_stream_fetcher.h
@@ -97,7 +97,7 @@ private:
 
     static QMutex m_instMutex;
     static QMap<QByteArray, VMaxStreamFetcher*> m_instances;
-    int m_sequence;
+    quint8 m_sequence;
     qint64 m_lastChannelTime[256];
     qint64 m_lastMediaTime;
     qint64 m_emptyPacketTime;
